feat(storybook): implement initial ui and storybook for notification center template
